pyproject.toml
[project]
name = "local-tree-rag"
version = "0.1.0"
description = "A local-first RAG system based on PageIndex reasoning"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "streamlit",
    "pymupdf",
    "openai",
    "groq",
    "python-dotenv",
    "tiktoken",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"